+++ CHANGELOG.md
@@ -8,3 +8,8 @@
 
 - Cleanup
 - Publish 0.1.0
+
+## 0.1.1
+
+- Add named optional parameter 'rootTag' to `convert`
+- Add named optional parameter 'imageBaseUrl' to `convert`